document-manger-api/DMS/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyhdb==0.3.4
pydantic
python-dotenv==1.0.0
requests
sap-xssec